- Fix assert (exclusive of 64 bits version) introduced in r12861.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@12864 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/SpellChecker/StatusField.cpp
@@ -101,7 +101,7 @@ static wxBitmapBundle LoadImageInPath(const wxString &path, wxString fileName, c
     if (bitmaps.empty())
         msg.Printf("Loading images: no images found in path '%s'!", path);
     else
-        msg.Printf("Loading images: %lu were successfully loaded!", bitmaps.size());
+        msg.Printf("Loading images: %lu were successfully loaded!", (unsigned long)bitmaps.size());
 
     Manager::Get()->GetLogManager()->DebugLog(msg);
     return wxBitmapBundle::FromBitmaps(bitmaps);
